script para insertar users y teams

+++ back/team/models.py
@@ -6,7 +6,7 @@ class Team(models.Model):
     name = models.CharField(max_length=100)
     draft = models.ForeignKey(Draft, on_delete=models.CASCADE)
     draft_user = models.ForeignKey(DraftUser, on_delete=models.CASCADE)
-    budget = models.FloatField()
+    budget = models.DecimalField(decimal_places=2, max_digits=100)
     
     def __str__(self):
         return self.name
